Update exercises 96 - 100 exercise 97 and 98 done

+++ CursoemVideo/Exercicios Mundo 3/ex096_funçao_que_calcula_area.py
@@ -18,6 +18,6 @@ def tit(msg):
 
 
 tit('Calculando área')
-a = int(input('Largura: (m) '))
-b = int(input('Comprimento: (m) '))
+a = float(input('Largura: (m) '))
+b = float(input('Comprimento: (m) '))
 área(a, b)
